[nextjs-x-chain] Minor improvements
