+++ apps/nextjs-x-chain/src/app/components/index.ts
@@ -1,0 +1,4 @@
+export * from './AccountBalance';
+export * from './CCTPTransfer';
+export * from './WalletConnection';
+export * from './WalletSelection';
