+++ apps/nextjs-x-chain/src/app/page.tsx
@@ -1,34 +1,18 @@
 "use client";
 
-import { useEffect, useState } from "react";
-import { useAutoConnect } from "@/components/AutoConnectProvider";
-import { DisplayValue, LabelValueGrid } from "@/components/LabelValueGrid";
 import { ThemeToggle } from "@/components/ThemeToggle";
-import { WalletSelector as ShadcnWalletSelector } from "@/components/WalletSelector";
 import { SingleSigner } from "@/components/transactionFlows/SingleSigner";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Label } from "@/components/ui/label";
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
-import { Switch } from "@/components/ui/switch";
-import { isMainnet } from "@/utils";
-import { Network } from "@aptos-labs/ts-sdk";
-import {
-  AccountInfo,
-  AdapterWallet,
-  AptosChangeNetworkOutput,
-  NetworkInfo,
-  OriginWalletDetails,
-  isAptosNetwork,
-  useWallet,
-} from "@aptos-labs/wallet-adapter-react";
-import { init as initTelegram } from "@telegram-apps/sdk";
-import { AlertCircle } from "lucide-react";
-import Image from "next/image";
 // Imports for registering a browser extension wallet plugin on page load
 import { MyWallet } from "@/utils/standardWallet";
+import { Network } from "@aptos-labs/ts-sdk";
+import { OriginWalletDetails, useWallet } from "@aptos-labs/wallet-adapter-react";
 import { registerWallet } from "@aptos-labs/wallet-standard";
-import { CCTPTransfer } from "@/components/CCTPTransfer";
+import { init as initTelegram } from "@telegram-apps/sdk";
+import { AlertCircle } from "lucide-react";
+import { useEffect, useState } from "react";
+
+import { AccountBalance, CCTPTransfer, WalletConnection, WalletSelection } from './components';
 
 // Example of how to register a browser extension wallet plugin.
 // Browser extension wallets should call registerWallet once on page load.
@@ -66,8 +50,8 @@ export default function Home() {
       const details = await getOriginWalletDetails(wallet);
       setOriginWalletDetails(details);
     };
-    fetchOriginWalletDetails();
-  }, [wallet]);
+    void fetchOriginWalletDetails();
+  }, [wallet, getOriginWalletDetails]);
 
   return (
     <main className="flex flex-col w-1/2 p-6 pb-12 md:px-8 gap-6">
@@ -89,252 +73,40 @@ export default function Home() {
         <ThemeToggle />
       </div>
       <WalletSelection />
-      {connected && (
-        <WalletConnection
-          account={account}
-          network={network}
-          wallet={wallet}
-          changeNetwork={changeNetwork}
-          originWalletDetails={originWalletDetails}
-        />
-      )}
-      {connected && isMainnet(connected, network?.name) && (
-        <Alert variant="warning">
-          <AlertCircle className="h-4 w-4" />
-          <AlertTitle>Warning</AlertTitle>
-          <AlertDescription>
-            The transactions flows below will not work on the Mainnet network.
-          </AlertDescription>
-        </Alert>
-      )}
-      {connected && (
-        <CCTPTransfer
-          wallet={wallet}
-          originWalletDetails={originWalletDetails}
-        />
-      )}
-      {connected && (
+
+      {connected && wallet && account && network && (
         <>
+          <WalletConnection
+            account={account}
+            network={network}
+            wallet={wallet}
+            changeNetwork={changeNetwork}
+            originWalletDetails={originWalletDetails}
+          />
+          {network?.name === Network.MAINNET && (
+            <Alert variant="warning">
+              <AlertCircle className="h-4 w-4" />
+              <AlertTitle>Warning</AlertTitle>
+              <AlertDescription>
+                The transactions flows below will not work on the Mainnet network.
+              </AlertDescription>
+            </Alert>
+          )}
+          <AccountBalance
+            account={account}
+            network={network}
+            wallet={wallet}
+          />
           <SingleSigner />
+          {network?.name !== Network.DEVNET && (
+            <CCTPTransfer
+              wallet={wallet}
+              originWalletDetails={originWalletDetails}
+            />
+          )}
         </>
       )}
     </main>
   );
 }
 
-function WalletSelection() {
-  const { autoConnect, setAutoConnect } = useAutoConnect();
-
-  return (
-    <Card>
-      <CardContent>
-        <div className="flex flex-wrap gap-6 pt-6 pb-12 justify-between items-center">
-          <div className="flex flex-col gap-4 items-center">
-            <ShadcnWalletSelector />
-          </div>
-        </div>
-        <label className="flex items-center gap-4 cursor-pointer">
-          <Switch
-            id="auto-connect-switch"
-            checked={autoConnect}
-            onCheckedChange={setAutoConnect}
-          />
-          <Label htmlFor="auto-connect-switch">
-            Auto reconnect on page load
-          </Label>
-        </label>
-      </CardContent>
-    </Card>
-  );
-}
-
-interface WalletConnectionProps {
-  account: AccountInfo | null;
-  network: NetworkInfo | null;
-  wallet: AdapterWallet | null;
-  originWalletDetails: OriginWalletDetails | undefined;
-  changeNetwork: (network: Network) => Promise<AptosChangeNetworkOutput>;
-}
-
-function WalletConnection({
-  account,
-  network,
-  wallet,
-  changeNetwork,
-  originWalletDetails,
-}: WalletConnectionProps) {
-  const { isSolanaDerivedWallet, isEIP1193DerivedWallet } = useWallet();
-
-  const isValidNetworkName = () => {
-    if (isAptosNetwork(network)) {
-      return Object.values<string | undefined>(Network).includes(network?.name);
-    }
-    // If the configured network is not an Aptos network, i.e is a custom network
-    // we resolve it as a valid network name
-    return true;
-  };
-
-  const isNetworkChangeSupported =
-    wallet?.features["aptos:changeNetwork"] !== undefined;
-
-  const aptosAccountInfoLabels = [
-    {
-      label: "Address",
-      value: (
-        <DisplayValue
-          value={account?.address?.toString() ?? "Not Present"}
-          isCorrect={!!account?.address}
-        />
-      ),
-    },
-    {
-      label: "Public key",
-      value: (
-        <DisplayValue
-          value={account?.publicKey?.toString() ?? "Not Present"}
-          isCorrect={!!account?.publicKey}
-        />
-      ),
-    },
-    {
-      label: "ANS name",
-      subLabel: "(only if attached)",
-      value: <p>{account?.ansName ?? "Not Present"}</p>,
-    },
-  ];
-  if (
-    wallet &&
-    (isSolanaDerivedWallet(wallet) || isEIP1193DerivedWallet(wallet))
-  ) {
-    aptosAccountInfoLabels.push({
-      label: "Origin Wallet Address",
-      value: (
-        <DisplayValue
-          value={originWalletDetails?.address.toString() ?? "Not Present"}
-          isCorrect={!!originWalletDetails?.address}
-        />
-      ),
-    });
-  }
-
-  return (
-    <Card>
-      <CardHeader>
-        <CardTitle>Wallet Connection</CardTitle>
-      </CardHeader>
-      <CardContent className="flex flex-col gap-10 pt-6">
-        <div className="flex flex-col gap-6">
-          <h4 className="text-lg font-medium">Wallet Details</h4>
-          <LabelValueGrid
-            items={[
-              {
-                label: "Icon",
-                value: wallet?.icon ? (
-                  <Image
-                    src={wallet.icon}
-                    alt={wallet.name}
-                    width={24}
-                    height={24}
-                  />
-                ) : (
-                  "Not Present"
-                ),
-              },
-              {
-                label: "Name",
-                value: <p>{wallet?.name ?? "Not Present"}</p>,
-              },
-              {
-                label: "URL",
-                value: wallet?.url ? (
-                  <a
-                    href={wallet.url}
-                    target="_blank"
-                    rel="noreferrer"
-                    className="text-blue-600 dark:text-blue-300"
-                  >
-                    {wallet.url}
-                  </a>
-                ) : (
-                  "Not Present"
-                ),
-              },
-            ]}
-          />
-        </div>
-
-        <div className="flex flex-col gap-6">
-          <h4 className="text-lg font-medium">Aptos Account Info</h4>
-          <LabelValueGrid items={aptosAccountInfoLabels} />
-        </div>
-
-        <div className="flex flex-col gap-6">
-          <h4 className="text-lg font-medium">Network Info</h4>
-          <LabelValueGrid
-            items={[
-              {
-                label: "Network name",
-                value: (
-                  <DisplayValue
-                    value={network?.name ?? "Not Present"}
-                    isCorrect={isValidNetworkName()}
-                    expected={Object.values<string>(Network).join(", ")}
-                  />
-                ),
-              },
-              {
-                label: "URL",
-                value: network?.url ? (
-                  <a
-                    href={network.url}
-                    target="_blank"
-                    rel="noreferrer"
-                    className="text-blue-600 dark:text-blue-300"
-                  >
-                    {network.url}
-                  </a>
-                ) : (
-                  "Not Present"
-                ),
-              },
-              {
-                label: "Chain ID",
-                value: <p>{network?.chainId ?? "Not Present"}</p>,
-              },
-            ]}
-          />
-        </div>
-
-        <div className="flex flex-col gap-6">
-          <h4 className="text-lg font-medium">Change Network</h4>
-          <RadioGroup
-            value={network?.name}
-            orientation="horizontal"
-            className="flex gap-6"
-            onValueChange={(value: Network) => changeNetwork(value)}
-            disabled={!isNetworkChangeSupported}
-          >
-            <div className="flex items-center space-x-2">
-              <RadioGroupItem value={Network.DEVNET} id="devnet-radio" />
-              <Label htmlFor="devnet-radio">Devnet</Label>
-            </div>
-            <div className="flex items-center space-x-2">
-              <RadioGroupItem value={Network.TESTNET} id="testnet-radio" />
-              <Label htmlFor="testnet-radio">Testnet</Label>
-            </div>
-            <div className="flex items-center space-x-2">
-              <RadioGroupItem value={Network.MAINNET} id="mainnet-radio" />
-              <Label htmlFor="mainnet-radio">Mainnet</Label>
-            </div>
-          </RadioGroup>
-          {!isNetworkChangeSupported && (
-            <div className="text-sm text-red-600 dark:text-red-400">
-              * {wallet?.name ?? "This wallet"} does not support network change
-              requests
-            </div>
-          )}
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
