+++ apps/nextjs-x-chain/src/components/WalletProvider.tsx
@@ -9,8 +9,9 @@ import { useClaimSecretKey } from "@/hooks/useClaimSecretKey";
 import { useAutoConnect } from "./AutoConnectProvider";
 import { useToast } from "./ui/use-toast";
 
-setupAutomaticEthereumWalletDerivation({ defaultNetwork: Network.TESTNET });
-setupAutomaticSolanaWalletDerivation({ defaultNetwork: Network.TESTNET });
+// Temporarily disabling this, as there's no on-chain support
+// setupAutomaticEthereumWalletDerivation({ defaultNetwork: Network.DEVNET });
+setupAutomaticSolanaWalletDerivation({ defaultNetwork: Network.DEVNET });
 
 let dappImageURI: string | undefined;
 if (typeof window !== "undefined") {
